backend/requirements.txt
Flask==3.0.3
pymongo==4.7.2
PyJWT==2.8.0
flask-cors==4.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-google-genai==1.0.8
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pypdf==4.2.0
python-dotenv==1.0.1
werkzeug==3.0.3
